requirements.txt
attrs==21.2.0
Automat==20.2.0
boto3==1.21.46
botocore==1.24.46
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
constantly==15.1.0
crayons==0.4.0
cryptography==3.4.7
cssselect==1.1.0
geckodriver-autoinstaller==0.1.0
greenlet==1.1.2
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itemadapter==0.3.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
numpy==1.21.1
pandas==1.3.1
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.1
queuelib==1.6.1
requests==2.26.0
s3transfer==0.5.2
Scrapy==2.4.0
scrapy-deltafetch==2.0.1
scrapy-selenium==0.0.7
scrapy-selenium-middleware==0.0.5
selenium==3.141.0
selenium-wire==2.1.1
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.25
Twisted==21.7.0
typing-extensions==3.10.0.0
urllib3==1.26.6
w3lib==1.22.0
webdriver-manager==3.4.2
zope.interface==5.4.0
